Add url-shoarter

+++ aws-python-rest-api-with-dynamodb/todos/delete.py
@@ -1,6 +1,8 @@
+import json
 import os
 
 import boto3
+from boto3.dynamodb.conditions import Key
 
 from utils.auth_decorator import token_required
 
@@ -12,8 +14,16 @@ def delete(event, context, username):
     table = dynamodb.Table(os.environ["DYNAMODB_TABLE"])
 
     # delete the todo from the database
-    table.delete_item(Key={"id": event["pathParameters"]["id"]})
-
+    result = table.scan(
+        FilterExpression=Key('id').eq(event["pathParameters"]["id"]) & Key('user').eq(username)
+    )
+    if result["Items"]:
+        table.delete_item(Key={"id": event["pathParameters"]["id"]})
+    else:
+        return {
+            'statusCode': 404,
+            'body': json.dumps({'message': 'Todo not found.'})
+        }
     # create a response
     response = {"statusCode": 200}
